+++ Sources/EmceeLib/Utils/RuntimeDumpTimeouts.swift
@@ -1,0 +1,7 @@
+import Foundation
+import Models
+
+public let testTimeoutConfigurationForRuntimeDump = TestTimeoutConfiguration(
+    singleTestMaximumDuration: 20,
+    testRunnerMaximumSilenceDuration: 20
+)